Began new class
Began to try to solve the problem by using for or while loops instead of the reversing of the words as it was previously programmed.
Also added the unit test class to the same file, it is currently commented out but it will be added back in when it is finished.
Also added a test file.