Finds most used word
WordCount.py now finds and prints the most used word. To complete the project it must find and return the 100 most used words.

I aslo created a new .txt document without all the text added by the editor and is purely the shakespeare text for an accurate count on words.