Began Tensorflow, enhanced WordCount.py
Began learning tensorflow, hence tensortest.py and the untitled Jupyter Notebook.

WorkCount.py now has a separate sort function that takes a dictionary and the size of the sorted list and returns the list with the size specified in the second argument that is the the top words used in the dictionary in order.